Extract trace writer chunk helpers
